wip: Moved types back into reducer files.
