+++ client/components/Channel.tsx
@@ -10,7 +10,7 @@ import '../assets/css/NoUiSlider.css'
 import '../assets/css/Channel.css'
 import * as IO from '../../shared/constants/SOCKET_IO_DISPATCHERS'
 import { IChannelReference, IFader } from '../../shared/faders-types'
-import { ISettings } from '../../shared/settings-types'
+import { ISettings } from '../../shared/reducers/settingsReducer'
 import { storeShowChanStrip } from '../../shared/actions/settingsActions'
 import { withTranslation } from 'react-i18next'
 import { VuLabelConversionType } from '../../shared/constants/MixerProtocolInterface'
