+++ client/components/ChannelMonitorOptions.tsx
@@ -5,7 +5,7 @@ import '../assets/css/ChannelMonitorOptions.css'
 import { Store } from 'redux'
 import { connect } from 'react-redux'
 import { storeShowMonitorOptions } from '../../shared/actions/settingsActions'
-import { ISettings } from '../../shared/settings-types'
+import { ISettings } from '../../shared/reducers/settingsReducer'
 import {
     SOCKET_SET_AUX_LEVEL,
     SOCKET_SET_FADER_MONITOR,
