+++ server/__tests__/indexReducer.spec.ts
@@ -1,8 +1,8 @@
-import indexReducer from '../reducers/indexReducer'
+import indexReducer from '../../shared/reducers/indexReducer'
 
 let fs = require('fs')
 const parsedEmptyStoreJSON = fs.readFileSync(
-    'server/__tests__/__mocks__/parsedEmptyStore.json'
+    '__tests__/__mocks__/parsedEmptyStore.json'
 )
 
 describe('Test initialize store', () => {
@@ -10,7 +10,7 @@ describe('Test initialize store', () => {
     it('should return the initial state of the whole Store', () => {
         // ** Uncomment to update initial settings state:
         // let data = indexReducer(undefined, {type: ''})
-        // fs.writeFileSync('server/__tests__/__mocks__/parsedEmptyStore-UPDATE.json', JSON.stringify(data))
+        // fs.writeFileSync('__tests__/__mocks__/parsedEmptyStore-UPDATE.json', JSON.stringify(data))
 
         expect(
             indexReducer(JSON.parse(parsedEmptyStoreJSON), { type: '' })
