+++ server/tsconfig.json
@@ -1,25 +1,17 @@
 {
     "compilerOptions": {
-        // Target latest version of ECMAScript.
         "target": "ESNEXT",
         "lib": ["ESNEXT", "dom"],
-        // Search under node_modules for non-relative imports.
         "moduleResolution": "node",
-        // Process & infer types from .js files.
         "allowJs": false,
-        // Enable strictest settings like strictNullChecks & noImplicitAny.
         "strict": false,
-        // Disallow features that require cross-file information for emit.
         "isolatedModules": false,
-        // Import non-ES modules as default imports.
         "esModuleInterop": true,
-        // Check for "any" while converting to TS this can be turned off.
         "noImplicitAny": true,
         "resolveJsonModule": true,
         "module": "commonjs",
-        "outDir": "../dist",
-        "typeRoots": ["@types", "node_modules/@types"],
+        "outDir": "../dist"
     },
     "include": ["../server", "../shared"],
-    "exclude": ["../**/node_modules/*", "../**/dist/*", "../**/*.spec.ts"]
+    "exclude": ["node_modules/*", "dist/*", "**/*.spec.ts"]
 }
